+++ rows.go
@@ -356,11 +356,7 @@ func (r *rows) Next(dest []driver.Value) error {
 			default:
 				b := C.dpiData_getBytes(d)
 				s := C.GoStringN(b.ptr, C.int(b.length))
-				if r.NumberAsString() {
-					dest[i] = s
-				} else {
-					dest[i] = Number(s)
-				}
+				dest[i] = s
 				if Log != nil {
 					Log("msg", "b", "i", i, "ptr", b.ptr, "length", b.length, "typ", col.NativeType, "int64", C.dpiData_getInt64(d), "dest", dest[i])
 				}