+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.14.0]
+### Changed
+- Make NumberAsString() the default.
+- Remove NumberAsString() and MagicConversion() Options.
+
 ## [0.13.3]
 ### Added
 - Dummy .go files to allow vendoring with "go mod vendor".